Remove @ensured_* cache variables.
When running different run lists per role, @ensured_* variables are not refreshed, and for example, roundsman thinks it already created /tmp/roundsman, but that's not the case, since we're contacting a different host now. Since the ensure operations are quick and idempotent, no need to go extra mile just to skip them.